auth hooks
