+++ app/src/hooks/index.ts
@@ -1,3 +1,5 @@
 export { useSayHello } from './use-say-hello.hook';
 export { useGetVersion } from './use-get-version.hook';
 export { useLogin } from './use-login.hook';
+export { useIsLoggedIn } from './use-logged-in.hook';
+export { useServerUrl } from './use-server-url.hook';
