+++ app/src/hooks/use-logged-in.hook.ts
@@ -1,0 +1,13 @@
+export const useIsLoggedIn = () => {
+  const isLoggedIn = async (): Promise<boolean> => {
+    const tokens = await chrome.storage.local.get([
+      'accessToken',
+      'refreshToken',
+    ]);
+
+    const { accessToken, refreshToken } = tokens;
+    return (accessToken ?? false) && (refreshToken ?? false);
+  };
+
+  return { isLoggedIn };
+};
