+++ app/src/hooks/use-logout.hook.ts
@@ -1,0 +1,15 @@
+import { useNavigationStore } from '../store';
+
+export const useLogout = () => {
+  const { navigateTo } = useNavigationStore();
+
+  const logout = async () => {
+    await chrome.storage.local.set({
+      accessToken: '',
+      refreshToken: '',
+    });
+    navigateTo('login');
+  };
+
+  return { logout };
+};
