+++ app/src/hooks/use-server-url.hook.ts
@@ -1,0 +1,7 @@
+import { useAuthStore } from '../store';
+
+export const useServerUrl = () => {
+  const serverUrl = useAuthStore((state) => state.serverUrl);
+  const setServerUrl = useAuthStore((state) => state.setServerUrl);
+  return { serverUrl, setServerUrl };
+};
