+++ app/src/store/index.ts
@@ -1,2 +1,2 @@
 export { useNavigationStore } from './navigation.store';
-export { authStore, useAuthStore, useLogout, useServerUrl } from './auth.store';
+export { authStore, useAuthStore } from './auth.store';
